+++ creme/stats/rolling_mode.py
@@ -48,15 +48,15 @@ class RollingMode(base.Univariate, utils.Window):
     """
 
     def __init__(self, window_size):
-        super().__init__(window_size)
+        super().__init__(size=window_size)
         self.counts = collections.defaultdict(int)
 
     @property
     def name(self):
-        return f'rolling_{self.window_size}_mode'
+        return f'rolling_{self.size}_mode'
 
     def update(self, x):
-        if len(self) >= self.window_size:
+        if len(self) >= self.size:
 
             # Subtract the counter of the last element
             first_in = self[0]